# Conflicts:
#	run/server.properties
#	src/main/java/kr/rtuserver/cash/RSCash.java
#	src/main/java/kr/rtuserver/cash/cash/CashManager.java
#	src/main/java/kr/rtuserver/cash/configuration/CashConfig.java
#	src/main/java/kr/rtuserver/cash/configuration/CoinConfig.java
#	src/main/java/kr/rtuserver/cash/dependency/PlaceholderAPI.java
#	src/main/java/kr/rtuserver/cash/listeners/CoinInteract.java
#	src/main/java/kr/rtuserver/cash/listeners/PlayerJoin.java

+++ src/main/java/kr/rtuserver/cash/cash/Coin.java
@@ -1,5 +1,5 @@
 package kr.rtuserver.cash.cash;
 
-public record Coin(String cash, String item, int value) {
+public record Coin(String cash, String item, long value) {
 
 }
