+++ src/main/java/kr/rtuserver/cash/cash/PlayerCash.java
@@ -1,5 +1,5 @@
 package kr.rtuserver.cash.cash;
 
-public record PlayerCash(String name, int cash) {
+public record PlayerCash(String name, long cash) {
 
 }
